hello_world/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
dnspython==2.6.1
fastapi==0.115.0
h11==0.14.0
idna==3.10
motor==3.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
